[STEL-37] FIX : Move moderator SCSS files to styles directory
Relocated all moderator-related SCSS files from 'pages/moderator' to 'styles/pages/moderator' and updated import paths in corresponding TypeScript files. This improves project structure by separating styles from component logic.
