+++ frontend/src/styles/pages/moderator/PollsModeration.scss
