+++ frontend/src/styles/pages/moderator/SpotsModeration.scss
